daala_tx: Force-inline SIMD functions.
GCC turns out to make bad choices about what to inline, e.g.,
reducing the 8x8 transpose to 2 live registers and passing
everything into and out of it on the stack (with essentially two
loads and a store for every operation). Forcing things to be
inline produces much better code.

Change-Id: If76d191e85d141492f8ce8c7e15d01354bbe8426