Update patch set 10
Patch Set 10:

Tom, I'm working with Andrey Norkin from Netflix on providing a reference implementation of the noise regraining in AV1 (in particular, the parameter estimation part that would live in the encoder). It may take some time before we get a full e2e estimation in the encoder (as that will require a denoiser as well).

Can you help me find an appropriate reviewer? I'm not entirely sure if this should be in aom_dsp, in the av1 folder, or in some other place reserved for examples.

There are a couple of areas that are still in need of improvement (e.g., chroma subsmapling is not handled correctly). But I'm sure there are plenty of other stylistic matters to get out of the way first.

Patch-set: 10
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Reviewer: Gerrit User 8612 <8612@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>