update filter mask calculation for 5 tap filter on chroma plane
when 5 tap filter(filter6) was introduced for chroma plane, it reads 6 pixel
lines/columns and modify 4 pixels. However, for the filter mask
calculation, it still use the original filter mask calculation function
for 7 tap filter(filter8), i.e, read 8 pixel lines/columns to check whether the
edge sould be filtered or not. This change updates the filter mask
calculation to make it consistent with the filering operation to only
read 6 pixels for filter mask calculation.

Change-Id: Iec434e48161e1bc2e67b38ea33c46e9c73774ce2