Remove global offset components from warped motion samples
To calculate wmmat[2] to wmmat[5], we don't need the absolute
coordinates which include the global offset components. Hence,
while finding samples, we only need to gather the relative coordinates
that are the neighbor block center point's coordinates relative to the
left-top pixel of current block. This patch won't change bitstream.

Change-Id: Ifbe7769effe6253b27cf97ea9c942a3a29eb7f50