Update patch set 6
Patch Set 6:

> Patch Set 5: Code-Review+2
> 
> Thanks for the effort - Yue and I once together worked on the following CL but we did not get to the correct default CDF which should be 16384, instead of 32768:
> https://aomedia-review.googlesource.com/c/aom/+/28083
> 
> Please fix the following typos in the comment before merging:
> (a) signle -> single
> (2) caculate -> calculate
> (3) gaim -> gain

The default cdf was not the problem, previously we changed them to 32767 to bypass the sanity check error report in cost calculation. The previous patch failed some unit tests for some other unknown reason, so it is great that Hui breaks the task into smaller patches and clear them progressively. And it would be exciting to have all the binary probs removed in the end. If I remember correctly, symbols related to reference frames are the only things left.

Patch-set: 6