[dist-8x8] Remove asserts for sum of sub8x8 partitions
Remove the asserts in rd_pick_partition(), which check
whether the sum of distortions of sub8x8 partitions matches to
the 8x8 distortion measured in final reconstructed 8x8 pixels,
because the sum of the sub8x8 distortions returned from rd_pick_sb_modes()
is sometimes different from what is measured with final reconstructed 8x8 pixels.

The asserts have been provided for the purpose of sanity check with PSNR
and triggered with "--enable-dist-8x8=1" encoder option.

Change-Id: I08e331900b19077cbb325ef5117688f05095dcc7