Fix txk-sel rd search in the key frames
If the transform block selects a transform kernel type that makes
it all zero coefficients and hits the best rd cost, store such
decision and force the transform block to be coded as all zero
block in the final coding pass. This makes txk-sel improve the
kf coding performance by 0.15% now.

Change-Id: I4bc3c185b9c68daefb856276da45bbf9e317d02e