Update patch set 25
Patch Set 25:

These looks like good changes to me.

Does it address the secondary issue in 1091?

i.e. if the resolution changes, is it legal for the previous segmentation map to be used?

If this is a legal case, I am worried that we may end up getting segmentation errors when accessing the segmentation map from a smaller reference frame.
If it is not a legal case, there should probably be an assert in the decoder to catch the error case.

Patch-set: 25
Reviewer: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>