Update patch set 9
Patch Set 9:

> Patch Set 6:
> 
> (1 comment)

It depends on what this monochrome video is. I added a separate flag to be able not to send the color info at all. It should be useful if the monochrome video is depth, for example.

Patch-set: 9
Reviewer: Gerrit User 6720 <6720@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>