Use SPLIT/HORZ/VERT to completely predict PART3
This patch is targeted to completely skip the rd pick mode stage, and
use the selected modes from SPLIT/HORZ/VERT for the 4 partitions of
HORZA, HORZB, VERTA, VERTB.

This patch has not been completed yet. In particular, it needs to figure
out how to obtain the RD cost from the existing mode.

Change-Id: I44070413e7b9802018ce8e4d762eb16d01266b01