Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> lgtm
> 
> just curious that the goal is to add synthetic noise towards the reconstructed frames eventually(?)

Yes, so far the noise synthesis code that I added is strictly for testing (and noise_util may be better off in the test folder). Andrey Norkin is working on the reference implementation for decoder.

In terms of adding end-to-end functionally in the encoder, I have some of the pieces ready but would need to discuss expectations of the denoiser if we choose to go that route (I have a simple Wiener 2D denoiser implemented).

Patch-set: 11
Reviewer: Gerrit User 8612 <8612@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>