Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+2
> 
> just curious, any impact on compression besides making things consistent?

The performance improvement on lowres is quite tiny, only 0.035%. This is why I hold on it. Yunqing has a patch that will fix interintra RDO a little bit. I will rebase this patch on top of that patch and test it in AWCY to see if the AWCY results can convince myself to check it this bitstream change.

Patch-set: 4
Reviewer: Gerrit User 5188 <5188@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>