Remove use_upsampled_ref speed feature.
The behavior of this feature was recently changed to do
upscaling/convolve on the fly as needed -- instead of upsampling the
references. So, the name was a misnomer.

Also, as this doesn't require additional memory now, we can always use
it -- irrespective of --cpu_used / frame resolution.

So, we remove this speed feature and the encoder now behaves as if that
feature was always on.

Change-Id: I242aebf07d60bca957cbff1db4fffde940beb00d