Remove resolution limit for using upsampled refs.
The behavior of this feature was recently changed to do
upscaling/convolve on the fly as needed -- instead of upsampling the
references. So, the name was a misnomer.

Also, as this doesn't require additional memory now, we can always use
it -- irrespective of --cpu_used / frame resolution.

Compression performance:
Overall: -0.15

Netflix_TunnelFlag:          -1.14
Netflix_Aerial:	             -0.66
touchdown_pass:              -0.49
Netflix_SquareAndTimelapse:  -0.40

https://arewecompressedyet.com/?job=with-tx64x64%402018-01-17T19%3A20%3A42.090Z&job=upsampled_ref_remove_limit%402018-01-18T19%3A36%3A53.025Z

Change-Id: Ic3d08fbff645121b2f95b279b8053cad130dab93