Added tests for horizontal superres
**DO NOT SUBMIT YET**--some changes still need to be made.

This patch adds end-to-end tests for horizontal superres and tests:

* 3 different videos with different bit depths and profiles, one of which
  contains screen content

* the 3 different superres modes

* a selection of values for superres_denominator and
  superres_kf_denominator

* a selection of values for superres_qthresh and superres_kf_qthresh
  (when superres_mode == SUPERRES_QTHRESH)

Note that tile_columns is set to 0 in this test, as non-zero values cause
seg faults (see Issue 1494). This value can be increased once the problem
has been resolved.

The tests are performed with target_bit_rate=40 and cpu_used=8 for speed.

BUG=aomedia:1390

Change-Id: Ic2fb52c7638241247dc6c4b232ca3450e67073cc