[NORMATIVE] Unify comp ref context design
This patch uses the neighboring ref counts to design the contexts
for the coding of the first reference frame of a reference pair for
the compound prediction. This will align the context design with
that for the second reference frame of a reference pair for the
compound prediction.

The default CDFs for the first reference frame coding have been
updated using aom_entropy_optimizer.

BUG=aomedia:1356

Change-Id: I781abbe4616dc3f3a7213ec663946ff9844eb830