[loop-restoration, bugfix] Restrict sampling of deblocked pixels
There is a special case with certain frame heights, where we
end up with a loop restoration stripe which ends 1px above the
crop border.

Previously this case was handled in quite an ugly way, which also
disagrees with the spec (+ isn't great for hardware). This patch
changes things to match the spec.

Specifically, the old method was to sometimes upscale one extra
row of deblocked pixels so that we could always have a 2px
"below" border for each processing stripe. The new method is to
only use rows inside the crop border, and to duplicate them if
necessary.

BUG=aomedia:1264

Change-Id: Idf8ab510e1091dc3f5b257de60e16bca214d8dc4