Update patch set 1
Patch Set 1:

Hi Deepa,

I am looking forward to a faster avx2 code for av1_convolve_2d_sr_avx2(),
which has been just landed yesterday as in:
https://aomedia.googlesource.com/aom/+/54cd8d76be6be0f77b459735d616d7fe21ae5630

Please rebase on it as well.

It would be appreciated if the speedup factors by this patch can be reported as well. Thanks!

Patch-set: 1
CC: Gerrit User 5035 <5035@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>