Fix the reference selection issue for skip mode
When the skip mode is used, it is possbile to select the two closest
forward references as the default reference pair. Such reference pair
may not be listed for the explicit signalling for the coding tool of
EXT_COMP_REFS.

This CL is to simply extend the reference pair candidate set for the
motion vector reference selection, hence the potential issue with
reference pair selection for skip mode is fixed without the need to
change the syntax.

Experimental results show no coding performance change regardless of
CONFIG_EXT_COMP_REFS on/off.

BUG=aomedia:1258

Change-Id: I7abe6afaaa69fa88bf66b2e327447a2ec17b63f0