Update patch set 9
Patch Set 9:

(7 comments)

> Patch Set 9: Verified+1
> 
> Build Successful 
> 
> https://build.aomedia.org/jenkins/job/libaom__commit_trigger/24465/ : SUCCESS

Yaowu/Tom:  This is continuation of patch 38841, which was inadvertently merged and then reverted.  Patch 8 of this CL is the patch "prior to" the code review comments, and patch 9 is the latest patch with the corrections.  Thank you.

Patch-set: 9
Reviewer: Gerrit User 7963 <7963@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>