Modified testbench showing tx overflow
Test bench for tx is modified to include test vectors with
maximum energy. This triggers the 64-bit clipping process found
in function half_btf() for the 64-point transform with 12-bit
source data, causing a mismatch with existing SIMD code.
SIMD code can be modified to deal with 64-bit overflow and
match C mode but it's somewhat convoluted (see #else portion of
code in half_btf_sse4_1()).

Change-Id: I928b98d236918b8a1561c63d6c8ec3452ec98399