Update patch set 15
Patch Set 15:

I think it is probably a mistake to try and get across all the changes to firstpass.c in one go especially given that some of them relate to or require changes in other areas.

My suggestion would be to first get over the changes that normalize the errors to per MB values. This should not change results (other than perhaps at the rounding level) once all the changes (to all relevant files) are correctly in place (i.e. changes to thresholds that depend on the error values that are now per mb not per frame).

This change underpins quite a few others as it makes comparisons between different image formats a bit more sane.

Then try and pull in and test individual changes such as the bug fix in regard to zoom detection. That way if one of the changes suddenly breaks things (perhaps because of changes to other files that have been missed) it will be easier to isolate.

Patch-set: 15
Tag: mailMessageId=<CALRks=SV23ADy=HfxNf04Ye6NjFKTf_ZM1u5Zheuc9swvvdkkA@mail.gmail.com>