Update patch set 16
Patch Set 16:

> Patch Set 15:
> 
> I think it is probably a mistake to try and get across all the changes to firstpass.c in one go especially given that some of them relate to or require changes in other areas.
> 
> My suggestion would be to first get over the changes that normalize the errors to per MB values. This should not change results (other than perhaps at the rounding level) once all the changes (to all relevant files) are correctly in place (i.e. changes to thresholds that depend on the error values that are now per mb not per frame).
> 
> This change underpins quite a few others as it makes comparisons between different image formats a bit more sane.
> 
> Then try and pull in and test individual changes such as the bug fix in regard to zoom detection. That way if one of the changes suddenly breaks things (perhaps because of changes to other files that have been missed) it will be easier to isolate.

Unfortunately I do not think the step by step approach will be easy to do. 

In a way, I like the current results for vbr, since the previous version pretty much ignored the target-bitrate values and the deviation from the target could be huge. This version is a lot better in tracking the target bitrate. The only problem is a few outlier videos: highway, bridge_far, bridge_close, which get a lot worse. But a majority of videos do get better.

Besides the q mode certainly gets a lot better overall, and everybody outside google and AWCY is testing the q-mode anyways. So it will be good to get the gains for q-mode anyways.

Patch-set: 16
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>