[wedge/compound-segment, normative] Remove more rounding
This reduces the overall rounding in the masked blend process -
the result is now equivalent to having a single round operation
at the end of the prediction process.

This increases the range of the intermediate values inside
aom_blend_a64_d32_mask() by 2 bits, but has no effect on the
ranges of any values outside that function.

BUG=aomedia:1292

Change-Id: I1010ed94c7d8db75bb3d8157c864c5527005725b