Update patch set 7
Patch Set 7:

Hi Max, it would be nice to be able to combine use of the config file with an explicit arg in the cmd line. This would be useful in automation scenarios where you largely want the same parameters but just want to change a couple of them explicitly.

Patch-set: 7