FILM GRAIN: Move film grain parameters for Altref frames
When a hidden (Altref) frame is later used as a "show existing frame"
the current code stores the film grain parameters in the "show existing
frame". However, this requires a second full pass of the frame
to apply the parameters. It is more efficient to have the option of
applying the parameters at the time the Altref frame is being decoded.
To enable this we have made the following changes:
- a hidden frame signals if it will later be used as a "show existing
frame" by sending a new flag "show_in_future"
- a hidden frame with the "show_in_future" flag set sends the film
grain parameters in the hidden frame rather than in the show existing
frame. The parameters can be decoded at the time the hidden frame is
decoded.

Change-Id: Ib24d13a5d2604b23917ab3417c37d5e5092433ac