Update patch set 17
Patch Set 17:

> Patch Set 17:
> 
> Hi Maxym and Erik,
> 
> I would like your pro tips on this patch. So I tried some complicated stuff like  _mm_i64gather_epi64 and _mm256_loadu2_m128i, but they only seemed to make the code slower. Let me know what you think. With this patch completed we know have have SIMD for the Low bit depth 4:2:0 CfL path.

I'm not surprised on the gather, it is a high latency instruction with a complicated uop flow. In truth, the only real advantage of gather instructions is when you need to load non-contiguous data to which require further SIMD operations. In those cases, a gathered load instruction is typically more efficient than a block of scalar code, depending on access patterns.

Great job vectorizing most of the lad cfl path!

Patch-set: 17