Update patch set 2
Patch Set 2:

> cylical refresh is an important tool to boost background clarity
 > for typical video conference use case. If it does not cause a bug
 > problem, we would prefer to leave it in place.
 > 
 > -Yaowu

I understand. I see that it is maintained in vp9 but failed to do so in aom, as it has caused at least one bug, for examples spatial_seg and encoder side desync with '--test-decode=" as belows:
https://bugs.chromium.org/p/aomedia/issues/detail?id=1262&q=owner%3Apaulwilkins%40google.com%20 (FIXED)
and
https://bugs.chromium.org/p/aomedia/issues/detail?id=1282&q=owner%3Apaulwilkins%40google.com%20&desc=3 (NOT FIXED YET)

I will remove all reviewers including you, since I need this patch for my branch.

Patch-set: 2