Scaled pred: proper convolve for subpel search
This patch deals with the case where a reference frame is not the same
resolution as the current frame.

Prior to this patch, aom_upsampled_pred() and aom_highbd_upsampled_pred()
used scaled (to the same resolution as the current frame) versions of the
reference frames to produce the inter predictions during subpel searches.
This patch instead generates the predictions by using a scaled convolve
on the unscaled reference frames, as this is closer to how the final
inter prediction is generated, and yields improved results.

In addition, PSNR thresholds in horz_superres_test.cc have been relaxed
slightly to ensure that the relevant tests pass.

Change-Id: Ie58be3f3ee39b2a2fa6f619f97d091acc1fe061e