Update patch set 3
Patch Set 3:

So, sub8x8_set_val() operates at TX dimensions and cfl_store_block() at the dimensions of visible TXs in a BLOCK.
In the case on 2:1 blocks, recording the last TX is sufficient because the top-left unit square covers the first TX of only 2.
There are 2 issues with 4:1 blocks:
If composed of 2x 2:1 TXs, the first TX is not completely covered by the top-left unit square.
If composed of 4x 1:1 TXs, the middle 2 are uncovered.
That said, the semantics of sub8x8_set_val() are unclear to me.

Patch-set: 3
Reviewer: Gerrit User 6340 <6340@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>