[convolve_rebalance_rounding] New experiment
The main purpose of this experiment is to make sure the output
of the interpolation filter fits in a 16bit range for 10bit content.

The second purpose is to reduce the rounding after the first (horizontal)
pass. The first pass is now rounded to Q4 rather than Q2.

Change round_0/round_1 from current 5/0 to 3/6. This increases precision
after first filter pass by 2 bits and decrease precision after second pass
by 4 bits. For bd=12, round_0 is kept at 5 to make sure the output of the
first filter pass still fit in 16bit.

Secondly, clamp the convolve output to the range

   [-0.5*(1<<bits), 1.5*(1<<bits)-1]
   where bits = bd + 14 - round_0 - round_1 = bd + 5.

This clamping is conservative as it still keep 50% overshoot margin in both
directions.

The AV1HighbdConvolve2DTest unit test is modified to ensure that the
worst case input is generated in order to trigger the clamping.
(without this change the clamping never triggers)

Change-Id: I2506a1d771b264aad0db0b3487c849e174b5373e