Update patch set 1
Patch Set 1: Code-Review-2

This check seems valid, and I think we should keep it -- to make sure the function is not called when scaling is off.

As mentioned in the referred bug, pls check on the tip-of-tree and reply if the bug still exists.

Patch-set: 1
Reviewer: Gerrit User 5290 <5290@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=-2