Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > > Patch Set 4:
> > > > 
> > > > (1 comment)
> > > 
> > > I assumed the optimiser would be packing the 8 bit cases into 64 bit writes, so this code is just easier to read.  Do you believe this isn't the case?
> > 
> > I am not sure if current generation of compilers can unroll the loop and pack the writes automatically, especially given there is a branch within the inner loop. 
> > 
> > The if-else branch itself should be moved outside for-loops too, from performance perspective.
> 
> Thanks for your response.  I don't have time to look into the optimisers in detail, but what you say surprises me.
> 
> Please do update this patch to look however you like.  The most important thing is to get something that is functionally correct checked in asap.

I just tried compiling the following:

int main() {

  const int size = 4096;
  char *dat = new char[size];

#if 1
  memset(dat, 5, size);
#else
  for (int i=0; i<4096; i++) {
    dat[i] = 5;
  }
#endif

  for (int j=0; j<size; j++) {
    std::cout << dat[j] << " ";
  }
  std::cout << std::endl;

  delete[] dat;
  return 0;
}

with g++ -O3 (both sides of that #if).
The output is identical both for the executable and the assembled code.

This should be the expected result: the compiler and optimiser should be working for the best for the architecture on which they are compiling.  Otherwise we'd all need to write hand-crafted code for each architecture (like we do with the SSE optimisations).

Here the compiler and optimiser should be doing the right thing to pack these 8 bit and 16 bit writes as efficiently as possible.

If you disagree, I'd be interested to see the evidence.

Patch-set: 4
Reviewer: Gerrit User 6145 <6145@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>