Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > I do not think we need sr_y to be changed. Instead we need jnt_convolve_y to be changed. I have updated your patch to revert the sr_y changes and added jnt_onvolve_y instead.
> 
> Hi Debargha
> 
> In my understanding the sr_y modification is needed. If not the C_Y/AV1Convolve2DSrTest fails when you set round_1!=0. This is because the sr_y special case code will no longer round the same way as the generic 2d-code (see below).
>  
> 
> [ RUN      ] C_Y/AV1Convolve2DSrTest.CheckOutput/0
> /work/olahug01/project/aom/test/av1_convolve_2d_test_util.cc:210: Failure
>       Expected: output[idx]
>       Which is: 'j' (106, 0x6A)
> To be equal to: output2[idx]
>       Which is: 'i' (105, 0x69)
> 4x4 Pixel mismatch at index 64 = (1, 0), sub pixel offset = (15, 0)
> [  FAILED  ] C_Y/AV1Convolve2DSrTest.CheckOutput/0, where GetParam() = (0xb347e6, 0, 1, 0) (0 ms)
> [----------] 1 test from C_Y/AV1Convolve2DSrTest (0 ms total)

Whenever we call the sr versions round_1 is 0 today. That is why I thought we do not need to change it. But I guess it is failing now because of the new tests. 
In that case, we can definitely bring it back, or change the test to remove the round_1 > 0 cases. I can reinstate those changes later today in the patch. Or if you get to it sooner - go ahead and change it back.

Patch-set: 3
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>