[NORMATIVE]ref_mv doesn't search different ref frame for smaller blocks
In ref_mv search, if less than 2 ref mv with same reference frame are
found, the mv with different reference frame will be searched and added
to mv_ref_list.

The CL modified the process. If the current block's width or height is
smaller or equal to 4, we don't search the difference reference frame
ref_mv, and just fills mv_ref_list with zeromv.

BUG=aomedia:1354

Change-Id: I846b469200f0c94273df66d301f30a8dbf4b2fa4