[non-normative, cfl] Fix build with cfl disabled
As of commit 2d276b39, reconintra.c cannot compile with CfL disabled,
since the UV_* versions of the intra modes are not defined.

Fix this by moving the UV_PREDICTION_MODE enum outside of its
enclosing #if CONFIG_CFL block. This requires a couple of minor
adjustments elsewhere in the code.

Change-Id: Id8d10ebfa627d8528719b557933929719d863a87