Update patch set 13
Patch Set 13:

(1 comment)

thanks for updating the change. for at least the speed tests it would be good to cover the C functions to get a baseline. if you're using C as the reference and not validating the output of that function on its own you can skip the other tests

Patch-set: 13
Reviewer: Gerrit User 5020 <5020@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>