Unify context design for single ref with comp ref
The CL makes the context design for single reference frame coding the
same as that for the compound reference frame coding. There are three
contexts designed for each of the binary symbols for the single
reference frame scenario, and the designed contexts simply rely on the
counts of the references used in the neighboring two blocks.

BUG=aomedia:1402

Change-Id: Ia72a1d18e85ac3a05308675b60b95f80f2219c46