Update patch set 1
Patch Set 1:

When an (always positive) loop number is a variable, not a constant, a for loop usually will have an extra comparison and jump to make sure loop number is positive. A do... while... loop can avoid this jump. That's why I changed them to this format. Considering this function is a heavy hitter, it's a trade-off between speed and readability.

Patch-set: 1
Reviewer: Gerrit User 5230 <5230@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>