Stop using VP9 convolve scheme in AV1 encoder.
Disable all VP9 style convolve rounding operations in the non-normative
parts of the encoder.

The function av1_convolve_2d_sr_c is forced instead on SIMD versions
of the same function, because of incompatibility.

In the -DCONFIG_LOWPRECISION_BLEND=2 -DCONFIG_HIGHPRECISION_INTBUF=1
setting, results on 15 frames of lowres (cpu-used=1) is -0.023% better.

Change-Id: I72154bd896357c352c944fb2cd3b25bafafba46a