[NORMATIVE] Add decoder support for frame_refs_signaling
One bit is added to the frame header, to signal whether the references
for inter frames will be short-signaled.

When the bit is set, only 6 bits are used at the frame header level to
signal the choices of LAST_FRAME and GOLDEN_FRAME out of the eight
buffered references. All the other 5 references will be specified
using the frame offset info. Hence, compared to the baseline, the use
of frame_refs_signaling provides 2 scenarios:

(1) Reference short-signaling:
    3*7 - (1+3*2) = 14 bits are saved in the frame header;
(2) Reference regular-signaling:
    1 bit is added to the frame header.

BUG=aomedia:1392

Change-Id: I2d4ff0e367a4df107235c1e7066dd28cb2c60431