[NORMATIVE] Fix top right mv check condition for VERT_A partition
When a coding block runs VERT_A partition, the coding order does
not follow raster order. This requires a special handle on the
bottom left square block to disable its reference towards the
top-right corner. Prior to this change the codebase would disable
the bottom-right square, as well as the right rectangular blocks,
from referencing the top-right mv. This commit fixes such check
condition to allow the right rectangular block to access the top
right mv.

BUG=aomedia:1343

Change-Id: I87049f0cec8ed7557a87c3fdef83e01498bbcd75