[NORMATIVE] Change lv_map to a single pass
This commit re-orders the syntax elements of the residual coding. It does not affect
coding efficiency since exactly the same symbols are coded, but in different order.

With this change the coefficients are parsed in a single pass. Previously the block
was required to be buffered between the golomb and the sign pass. Now coefficients
can be emitted on the fly.

Furthermore, signs are collected in 4-groups so that they can be read using an
unadapted multisymbol CDF where all outcomes have equal likelihood. This reduces
the number of entropy coding calls.

Previous syntax elemement order:
   for(coeffs)
       read base
       read br
   for(coeffs)
       read golomb
   for(coeffs)
       read sign bit

Changed to:
   for(coeffs)
       loop until 4 non-zero coeffs are parsed
           read base
           read br
       read up to 4 signs using a multisymbol
       for(up to 4)
           read golomb

The overall decoder is sped up by 1.1% when decoding city_cif.y4m at qp 20, ~1Mbit/s

Change-Id: I3e5e9dba5f78f62d8aa16eed3c0c9a599dc15d4c