[CFL] AVX2/SSSE3 faster luma_subsampling_420_lbd
Taking advantage of width and height being constant
propagated to speed up the SSSE3 implementation.
This new implementation is about twice the speed of
the previous version.

Based on the observation that for smaller widths,
the AVX2 code tends to be slower than its SSSE3
counterpart, the AVX2 code now calls the SSSE3
functions when width > 32.

SSSE3/CFLSubsampleTest
4x4: C time = 120 us, SIMD time = 80 us (~1.5x)
8x8: C time = 346 us, SIMD time = 96 us (~3.6x)
16x16: C time = 1327 us, SIMD time = 155 us (~8.6x)
32x32: C time = 5721 us, SIMD time = 530 us (~11x)

AVX2/CFLSubsampleTest
4x4: C time = 132 us, SIMD time = 89 us (~1.5x)
8x8: C time = 353 us, SIMD time = 111 us (~3.2x)
16x16: C time = 1290 us, SIMD time = 145 us (~8.9x)
32x32: C time = 5665 us, SIMD time = 524 us (~11x)

Change-Id: I9b2516ff5fc8addb2064904b6b362c8d89909933