Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi, Johann, megathanks for your input.
> 
> Right you are - I will try the simple patch first without sse4 stuff - just to be sure everything works as expected. It is my first time with Gerrit, sorry. I don't even know whether my reply here is the right place...
> 
> But you are not right about that  "

Looks like replies to the email almost work, but get cut at the "

I'm not sure what the minimum fallback is going to be for AOM. sse4 is probably reasonable. When writing new code, that's the best path. For existing code it's a little less clear. You should bring it up on the kavi codec mailing list.

In this instance however, you moved the sse2 code to a file name _sse4, but the functions are still named _sse2. With assembly, this might work. With intrinsics, the compiler will break this.

Patch-set: 1