Update patch set 7
Patch Set 7: Code-Review+1

(1 comment)

I am still worried that the encoder side of this may be wrong for the scalability case.

I would expect the encoder code to check that only existing film grain parameters from the same or lower layers were used.

There is an examples/scalable_decoder.c program that decodes scalable sequences, but as far as I understand it works by decoding all layers so I am not convinced that it would detect any failures.

On the other hand, I suspect that the whole code may have a lot of problems with scalability completely unrelated to this patch so perhaps it is best to address scalability issues at a different time and submit this patch as it is?

Patch-set: 7
Reviewer: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+1